Add route name in header for small screen sizes
